+++ moco-runner/src/main/java/com/github/dreamhead/moco/parser/deserializer/AbstractTextContainerDeserializer.java
@@ -17,7 +17,13 @@ import static com.google.common.collect.ImmutableMap.copyOf;
 import static com.google.common.collect.Maps.transformEntries;
 
 public abstract class AbstractTextContainerDeserializer<T extends TextContainer> extends JsonDeserializer<T> {
-    private final ImmutableMap<String, String> names = ImmutableMap.of("json_path", "jsonPaths", "xpath", "xpaths");
+    private final ImmutableMap<String, String> names = ImmutableMap.<String, String>builder()
+        .put("json_path", "jsonPaths")
+        .put("xpath", "xpaths")
+        .put("header", "headers")
+        .put("cookie", "cookies")
+        .put("form", "form")
+        .build();
 
     protected TextContainer textContainer(JsonParser jp, DeserializationContext ctxt) throws IOException {
         JsonToken currentToken = jp.getCurrentToken();